split the code into more files

+++ Types.hpp
@@ -2,5 +2,5 @@
 
 #include <complex>
 
-typedef long double real_t;
+typedef double real_t;
 typedef std::complex <real_t> complex_t;
